[FEAT] 서버에서 가져온 음악 리스트 가시화
- 서버에서 음악 가져오는 Extension 추가
- 음악 관련 Protocol 추가
- 음악 관련 코드 수정중
