+++ Whistle/Models/Music/MusicNote.swift
@@ -1,5 +1,5 @@
 //
-//  AudioPreview.swift
+//  AudioNote.swift
 //  Whistle
 //
 //  Created by 박상원 on 2023/09/05.
@@ -8,7 +8,7 @@
 import Foundation
 import SwiftUI
 
-struct AudioPreviewModel: Hashable {
+struct MusicNote: Hashable {
   var index: Int
   var magnitude: Float
   var color: Color
